rename exception.js, fix timeout
