+++ lib/gui/app.js
@@ -44,7 +44,7 @@ const availableDrives = require('./models/available-drives');
 const selectionState = require('./models/selection-state');
 const driveScanner = require('./modules/drive-scanner');
 const osDialog = require('./os/dialog');
-const exceptionReporter = require('./modules/exception');
+const exceptionReporter = require('./modules/exception-reporter');
 
 const app = angular.module('Etcher', [
   require('angular-ui-router'),
