+++ lib/gui/modules/exception-reporter.js
