+++ lib/gui/pages/main/controllers/drive-selection.js
@@ -19,7 +19,7 @@
 const settings = require('../../../models/settings');
 const selectionState = require('../../../models/selection-state');
 const analytics = require('../../../modules/analytics');
-const exceptionReporter = require('../../../modules/exception');
+const exceptionReporter = require('../../../modules/exception-reporter');
 
 module.exports = function(DriveSelectorService) {
