+++ lib/gui/pages/main/controllers/image-selection.js
@@ -26,7 +26,7 @@ const supportedFormats = require('../../../../shared/supported-formats');
 const analytics = require('../../../modules/analytics');
 const selectionState = require('../../../models/selection-state');
 const osDialog = require('../../../os/dialog');
-const exceptionReporter = require('../../../modules/exception');
+const exceptionReporter = require('../../../modules/exception-reporter');
 
 module.exports = function(
   $timeout,
@@ -132,22 +132,24 @@ module.exports = function(
    * ImageSelectionController.selectImageByPath('path/to/image.img');
    */
   this.selectImageByPath = (imagePath) => {
-    $timeout(() => {
-      imageStream.getImageMetadata(imagePath)
-        .then(this.selectImage)
-        .catch((error) => {
-          const imageError = errors.createUserError({
-            title: 'Error opening image',
-            description: messages.error.openImage({
-              imageBasename: path.basename(imagePath),
-              errorMessage: error.message
-            })
-          });
-
-          osDialog.showError(imageError);
-          analytics.logException(error);
+    imageStream.getImageMetadata(imagePath)
+      .then((imageMetadata) => {
+        $timeout(() => {
+          this.selectImage(imageMetadata);
         });
-    });
+      })
+      .catch((error) => {
+        const imageError = errors.createUserError({
+          title: 'Error opening image',
+          description: messages.error.openImage({
+            imageBasename: path.basename(imagePath),
+            errorMessage: error.message
+          })
+        });
+
+        osDialog.showError(imageError);
+        analytics.logException(error);
+      });
   };
 
   /**
