+++ lib/gui/pages/main/controllers/main.js
@@ -19,7 +19,7 @@
 const settings = require('../../../models/settings');
 const flashState = require('../../../models/flash-state');
 const analytics = require('../../../modules/analytics');
-const exceptionReporter = require('../../../modules/exception');
+const exceptionReporter = require('../../../modules/exception-reporter');
 const availableDrives = require('../../../models/available-drives');
 const selectionState = require('../../../models/selection-state');
