+++ lib/gui/pages/settings/controllers/settings.js
@@ -20,7 +20,7 @@ const os = require('os');
 const _ = require('lodash');
 const settings = require('../../../models/settings');
 const analytics = require('../../../modules/analytics');
-const exceptionReporter = require('../../../modules/exception');
+const exceptionReporter = require('../../../modules/exception-reporter');
 
 module.exports = function(WarningModalService) {
